requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.3
django-advanced-filters==1.3.0
django-braces==1.14.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
kombu==5.1.0
leaflet==0.0.3
prompt-toolkit==3.0.18
PyJWT==2.1.0
python-dotenv==0.17.1
pytz==2021.1
sentry-sdk==1.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
